+++ stress-jpeg.c
@@ -63,7 +63,7 @@ static int stress_set_jpeg_height(const char *opt)
 	int32_t jpeg_height;
 
 	jpeg_height = stress_get_int32(opt);
-	stress_check_range("jpeg-height", (uint32_t)jpeg_height, 256, 4096);
+	stress_check_range("jpeg-height", (uint64_t)jpeg_height, 256, 4096);
 	return stress_set_setting("jpeg-height", TYPE_ID_INT32, &jpeg_height);
 }
 
@@ -76,7 +76,7 @@ static int stress_set_jpeg_width(const char *opt)
 	int32_t jpeg_width;
 
 	jpeg_width = stress_get_int32(opt);
-	stress_check_range("jpeg-width", (uint32_t)jpeg_width, 256, 4096);
+	stress_check_range("jpeg-width", (uint64_t)jpeg_width, 256, 4096);
 	return stress_set_setting("jpeg-width", TYPE_ID_INT32, &jpeg_width);
 }
 
@@ -89,7 +89,7 @@ static int stress_set_jpeg_quality(const char *opt)
 	int32_t jpeg_quality;
 
 	jpeg_quality = stress_get_int32(opt);
-	stress_check_range("jpeg-quality", (uint32_t)jpeg_quality, 1, 100);
+	stress_check_range("jpeg-quality", (uint64_t)jpeg_quality, 1, 100);
 	return stress_set_setting("jpeg-quality", TYPE_ID_INT32, &jpeg_quality);
 }
